backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
psycopg2-binary==2.9.10
gunicorn==23.0.0
openai==1.82.0
python-dotenv==1.1.0
